requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
black==26.3.1
bleach==6.3.0
-e file:///home/saber/internships/assignments/cbct
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
connected-components-3d==3.26.1
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
fonttools==4.62.1
fqdn==1.5.1
fsspec==2026.2.0
h11==0.16.0
hatchling==1.29.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio==2.37.3
importlib-resources==6.5.2
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itk==5.4.5
itk-core==5.4.5
itk-filtering==5.4.5
itk-io==5.4.5
itk-numerics==5.4.5
itk-registration==5.4.5
itk-segmentation==5.4.5
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-client==8.8.0
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.4
jupyterlab==4.5.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
kiwisolver==1.5.0
lark==1.3.1
lazy-loader==0.5
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
monai==1.5.2
mpmath==1.3.0
mypy-extensions==1.1.0
narwhals==2.18.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nibabel==5.4.2
notebook==7.5.5
notebook-shim==0.2.4
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.13.0.92
overrides==7.7.0
packaging==26.0
pandas==3.0.1
pandocfilters==1.5.1
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
prometheus-client==0.24.1
prompt-toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydicom==3.0.1
pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytokens==0.4.1
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
ruff==0.15.6
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
send2trash==2.1.0
setuptools==82.0.1
simpleitk==2.5.3
simsimd==6.5.16
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
stringzilla==4.6.0
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2026.3.3
tinycss2==1.4.0
torch==2.10.0
torchaudio==2.5.1+cu121
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
triton==3.6.0
trove-classifiers==2026.1.14.14
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
